rend: Instance batching (#71)
* rend: Move instance comp to rend lib

* rend: Draw all renderables

* geo: Fix flipped perspective projection

* rend: Initial instance data api

* rend: Update naming on resource comps

* rend: Skip empty batches

* rend: Support combining instances

+++ libs/rend/src/rvk/uniform.c
@@ -101,7 +101,7 @@ void rvk_uniform_pool_destroy(RvkUniformPool* uni) {
   alloc_free_t(g_alloc_heap, uni);
 }
 
-usize rvk_uniform_size_max(RvkUniformPool* uni) { return uni->dataSizeMax; }
+u32 rvk_uniform_size_max(RvkUniformPool* uni) { return uni->dataSizeMax; }
 
 VkDescriptorSetLayout rvk_uniform_vkdesclayout(RvkUniformPool* uni) {
   return rvk_desc_vklayout(uni->device->descPool, &uni->descMeta);
